Handle retention cleanup TODOs. Minor reorg
