+++ relayer/processor/path_processor.go
@@ -97,31 +97,6 @@ func (pathEnd *pathEndRuntime) MergeCacheData(d ChainProcessorCacheData) {
 	pathEnd.ibcHeaderCache.Prune(ibcHeadersToCache) // Only keep most recent IBC headers
 }
 
-// packetIBCMessage holds a packet message's action and sequence along with it,
-// useful for sending packets around internal to the PathProcessor.
-type packetIBCMessage struct {
-	channelKey ChannelKey
-	action     string
-	sequence   uint64
-	message    provider.RelayerMessage
-}
-
-// channelIBCMessage holds a channel handshake message's action and channel along with it,
-// useful for sending messages around internal to the PathProcessor.
-type channelIBCMessage struct {
-	action     string
-	channelKey ChannelKey
-	message    provider.RelayerMessage
-}
-
-// connectionIBCMessage holds a connection handshake message's action and connection along with it,
-// useful for sending messages around internal to the PathProcessor.
-type connectionIBCMessage struct {
-	action        string
-	connectionKey ConnectionKey
-	message       provider.RelayerMessage
-}
-
 func NewPathProcessor(log *zap.Logger, pathEnd1 PathEnd, pathEnd2 PathEnd) *PathProcessor {
 	return &PathProcessor{
 		log: log,
@@ -258,115 +233,6 @@ func (pp *PathProcessor) HandleNewData(chainID string, cacheData ChainProcessorC
 	} else if pp.pathEnd2.info.ChainID == chainID {
 		pp.pathEnd2.incomingCacheData <- cacheData
 	}
-}
-
-// inProgressSend tracks the send state of an IBC message
-type inProgressSend struct {
-	sendHeight uint64
-	retryCount uint64
-	message    provider.RelayerMessage
-}
-
-type packetSendCache map[ChannelKey]packetChannelMessageCache
-type packetChannelMessageCache map[string]packetMessageSendCache
-type packetMessageSendCache map[uint64]inProgressSend
-
-func (c packetChannelMessageCache) deleteCachedMessages(toDelete ...map[string][]uint64) {
-	for _, toDeleteMap := range toDelete {
-		for message, toDeleteMessages := range toDeleteMap {
-			for _, sequence := range toDeleteMessages {
-				delete(c[message], sequence)
-			}
-		}
-	}
-}
-
-type channelSendCache map[string]channelKeySendCache
-type channelKeySendCache map[ChannelKey]inProgressSend
-
-func (c channelSendCache) deleteCachedMessages(toDelete ...map[string][]ChannelKey) {
-	for _, toDeleteMap := range toDelete {
-		for message, toDeleteMessages := range toDeleteMap {
-			for _, channel := range toDeleteMessages {
-				delete(c[message], channel)
-			}
-		}
-	}
-}
-
-type connectionSendCache map[string]connectionKeySendCache
-type connectionKeySendCache map[ConnectionKey]inProgressSend
-
-func (c connectionSendCache) deleteCachedMessages(toDelete ...map[string][]ConnectionKey) {
-	for _, toDeleteMap := range toDelete {
-		for message, toDeleteMessages := range toDeleteMap {
-			for _, connection := range toDeleteMessages {
-				delete(c[message], connection)
-			}
-		}
-	}
-}
-
-// contains MsgRecvPacket from counterparty
-// entire packet flow
-type pathEndPacketFlowMessages struct {
-	Src                   *pathEndRuntime
-	Dst                   *pathEndRuntime
-	ChannelKey            ChannelKey
-	SrcMsgTransfer        PacketSequenceCache
-	DstMsgRecvPacket      PacketSequenceCache
-	SrcMsgAcknowledgement PacketSequenceCache
-	SrcMsgTimeout         PacketSequenceCache
-	SrcMsgTimeoutOnClose  PacketSequenceCache
-}
-
-type pathEndConnectionHandshakeMessages struct {
-	Src                         *pathEndRuntime
-	Dst                         *pathEndRuntime
-	SrcMsgConnectionOpenInit    ConnectionMessageCache
-	DstMsgConnectionOpenTry     ConnectionMessageCache
-	SrcMsgConnectionOpenAck     ConnectionMessageCache
-	DstMsgConnectionOpenConfirm ConnectionMessageCache
-}
-
-type pathEndChannelHandshakeMessages struct {
-	Src                      *pathEndRuntime
-	Dst                      *pathEndRuntime
-	SrcMsgChannelOpenInit    ChannelMessageCache
-	DstMsgChannelOpenTry     ChannelMessageCache
-	SrcMsgChannelOpenAck     ChannelMessageCache
-	DstMsgChannelOpenConfirm ChannelMessageCache
-}
-
-type pathEndChannelCloseMessages struct {
-	Src                       *pathEndRuntime
-	Dst                       *pathEndRuntime
-	SrcMsgChannelCloseInit    ChannelMessageCache
-	DstMsgChannelCloseConfirm ChannelMessageCache
-}
-
-type pathEndPacketFlowResponse struct {
-	SrcMessages []packetIBCMessage
-	DstMessages []packetIBCMessage
-
-	ToDeleteSrc map[string][]uint64
-	ToDeleteDst map[string][]uint64
-}
-
-type pathEndChannelHandshakeResponse struct {
-	SrcMessages []channelIBCMessage
-	DstMessages []channelIBCMessage
-
-	ToDeleteSrc map[string][]ChannelKey
-	ToDeleteDst map[string][]ChannelKey
-}
-
-type pathEndConnectionHandshakeResponse struct {
-	SrcMessages []connectionIBCMessage
-	DstMessages []connectionIBCMessage
-
-	ToDeleteSrc map[string][]ConnectionKey
-	ToDeleteDst map[string][]ConnectionKey
 }
 
 // assembleIBCMessage constructs the applicable IBC message using the requested function.
@@ -648,6 +514,8 @@ ChannelHandshakeLoop:
 	}
 }
 
+// assembleMsgUpdateClient uses the ChainProvider from both pathEnds to assemble the client update header
+// from the source and then assemble the update client message in the correct format for the destination.
 func (pp *PathProcessor) assembleMsgUpdateClient(src, dst *pathEndRuntime) (provider.RelayerMessage, error) {
 	// If the client state trusted height is not equal to the client trusted state height, we cannot send a MsgUpdateClient
 	// until another block is observed. But we can send the rest without the MsgUpdateClient.
@@ -669,207 +537,7 @@ func (pp *PathProcessor) assembleMsgUpdateClient(src, dst *pathEndRuntime) (prov
 	return msgUpdateClient, nil
 }
 
-func (pathEnd *pathEndRuntime) shouldSendPacketMessage(message packetIBCMessage) bool {
-	msgSendCache, ok := pathEnd.packetSendCache[message.channelKey]
-	if !ok {
-		// in progress cache does not exist for this channel, so can send.
-		return true
-	}
-	channelSendCache, ok := msgSendCache[message.action]
-	if !ok {
-		// in progress cache does not exist for this action, so can send
-		return true
-	}
-	inProgress, ok := channelSendCache[message.sequence]
-	if !ok {
-		// in progress cache does not exist for this sequence, so can send.
-		return true
-	}
-	if pathEnd.latestBlock.Height-inProgress.sendHeight < blocksToRetryPacketAfter {
-		// this message was sent less than blocksToRetryPacketAfter ago, do not send again yet.
-		return false
-	}
-	if inProgress.retryCount == maxMessageSendRetries {
-		// giving up on sending this packet message
-		// TODO remove all retention of this sequence in pathEnd.messagesCache.PacketFlow and counterparty
-		return false
-	}
-
-	return true
-}
-
-func (pathEnd *pathEndRuntime) shouldSendConnectionMessage(message connectionIBCMessage) bool {
-	msgSendCache, ok := pathEnd.connectionSendCache[message.action]
-	if !ok {
-		// in progress cache does not exist for this action, so can send.
-		return true
-	}
-	inProgress, ok := msgSendCache[message.connectionKey]
-	if !ok {
-		// in progress cache does not exist for this connection, so can send.
-		return true
-	}
-	if pathEnd.latestBlock.Height-inProgress.sendHeight < blocksToRetryPacketAfter {
-		// this message was sent less than blocksToRetryPacketAfter ago, do not send again yet.
-		return false
-	}
-	if inProgress.retryCount == maxMessageSendRetries {
-		// giving up on sending this packet message
-		// TODO remove all retention of this connection handshake in pathEnd.messagesCache.ConnectionHandshake and counterparty
-		return false
-	}
-
-	return true
-}
-
-func (pathEnd *pathEndRuntime) shouldSendChannelMessage(message channelIBCMessage) bool {
-	msgSendCache, ok := pathEnd.channelSendCache[message.action]
-	if !ok {
-		// in progress cache does not exist for this action, so can send.
-		return true
-	}
-	inProgress, ok := msgSendCache[message.channelKey]
-	if !ok {
-		// in progress cache does not exist for this channel, so can send.
-		return true
-	}
-	if pathEnd.latestBlock.Height-inProgress.sendHeight < blocksToRetryPacketAfter {
-		// this message was sent less than blocksToRetryPacketAfter ago, do not send again yet.
-		return false
-	}
-	if inProgress.retryCount == maxMessageSendRetries {
-		// giving up on sending this packet message
-		// TODO remove all retention of this channel handshake in pathEnd.messagesCache.ChannelHandshake and counterparty
-		return false
-	}
-
-	return true
-}
-
-func (pathEnd *pathEndRuntime) trackSentPacketMessage(message packetIBCMessage) {
-	msgSendCache, ok := pathEnd.packetSendCache[message.channelKey]
-	if !ok {
-		pathEnd.packetSendCache[message.channelKey] = make(packetChannelMessageCache)
-	}
-	channelSendCache, ok := msgSendCache[message.action]
-	if !ok {
-		msgSendCache[message.action] = make(packetMessageSendCache)
-	}
-
-	retryCount := uint64(0)
-
-	if inProgress, ok := channelSendCache[message.sequence]; ok {
-		retryCount = inProgress.retryCount + 1
-	}
-
-	channelSendCache[message.sequence] = inProgressSend{
-		sendHeight: pathEnd.latestBlock.Height,
-		message:    message.message,
-		retryCount: retryCount,
-	}
-}
-
-func (pathEnd *pathEndRuntime) trackSentConnectionMessage(message connectionIBCMessage) {
-	msgSendCache, ok := pathEnd.connectionSendCache[message.action]
-	if !ok {
-		pathEnd.connectionSendCache[message.action] = make(connectionKeySendCache)
-	}
-
-	retryCount := uint64(0)
-
-	if inProgress, ok := msgSendCache[message.connectionKey]; ok {
-		retryCount = inProgress.retryCount + 1
-	}
-
-	msgSendCache[message.connectionKey] = inProgressSend{
-		sendHeight: pathEnd.latestBlock.Height,
-		message:    message.message,
-		retryCount: retryCount,
-	}
-}
-
-func (pathEnd *pathEndRuntime) trackSentChannelMessage(message channelIBCMessage) {
-	msgSendCache, ok := pathEnd.channelSendCache[message.action]
-	if !ok {
-		pathEnd.channelSendCache[message.action] = make(channelKeySendCache)
-	}
-
-	retryCount := uint64(0)
-
-	if inProgress, ok := msgSendCache[message.channelKey]; ok {
-		retryCount = inProgress.retryCount + 1
-	}
-
-	msgSendCache[message.channelKey] = inProgressSend{
-		sendHeight: pathEnd.latestBlock.Height,
-		message:    message.message,
-		retryCount: retryCount,
-	}
-}
-
-func (pp *PathProcessor) sendMessages(
-	ctx context.Context,
-	src, dst *pathEndRuntime,
-	packetMessages []packetIBCMessage,
-	connectionMessages []connectionIBCMessage,
-	channelMessages []channelIBCMessage,
-) error {
-	if len(packetMessages) == 0 && len(connectionMessages) == 0 && len(channelMessages) == 0 {
-		return nil
-	}
-
-	msgUpdateClient, err := pp.assembleMsgUpdateClient(src, dst)
-	if err != nil {
-		return fmt.Errorf("error prepending MsgUpdateClient, chain_id: %s, client_id: %s, %w",
-			dst.info.ChainID,
-			dst.info.ClientID,
-			err,
-		)
-	}
-
-	messages := []provider.RelayerMessage{msgUpdateClient}
-
-	for _, msg := range packetMessages {
-		if !dst.shouldSendPacketMessage(msg) {
-			continue
-		}
-		messages = append(messages, msg.message)
-	}
-	for _, msg := range connectionMessages {
-		if !dst.shouldSendConnectionMessage(msg) {
-			continue
-		}
-		messages = append(messages, msg.message)
-	}
-	for _, msg := range channelMessages {
-		if !dst.shouldSendChannelMessage(msg) {
-			continue
-		}
-		messages = append(messages, msg.message)
-	}
-
-	_, txSuccess, err := dst.chainProvider.SendMessages(ctx, messages)
-	if err != nil {
-		return fmt.Errorf("error sending messages to chain_id: %s, %v", dst.info.ChainID, err)
-	}
-	if !txSuccess {
-		return fmt.Errorf("error sending messages to chain_id: %s, transaction was not successful", dst.info.ChainID)
-	}
-
-	for _, msg := range packetMessages {
-		dst.trackSentPacketMessage(msg)
-	}
-	for _, msg := range connectionMessages {
-		dst.trackSentConnectionMessage(msg)
-	}
-	for _, msg := range channelMessages {
-		dst.trackSentChannelMessage(msg)
-	}
-
-	return nil
-}
-
-// updateClientTrustedState will combine the counterparty chains trusted IBC header
+// updateClientTrustedState combines the counterparty chains trusted IBC header
 // with the latest client state, which will be used for constructing MsgUpdateClient messages.
 func (pp *PathProcessor) updateClientTrustedState(src *pathEndRuntime, dst *pathEndRuntime) {
 	if src.clientTrustedState.ClientState.ConsensusHeight.GTE(src.clientState.ConsensusHeight) {
